+++ src/main/java/com/teethcare/common/Message.java
@@ -1,0 +1,5 @@
+package com.teethcare.common;
+
+public enum Message {
+    SUCCESS_FUNCTION
+}
