+++ src/main/java/com/teethcare/model/response/ErrorResponse.java
@@ -1,15 +1,15 @@
-package com.teethcare.model.dto;
+package com.teethcare.model.response;
 
-public class ErrorResponseDTO {
+public class ErrorResponse {
     private int status;
     private String message;
     private long timeStamp;
 
-    public ErrorResponseDTO() {
+    public ErrorResponse() {
 
     }
 
-    public ErrorResponseDTO(int status, String message, long timeStamp) {
+    public ErrorResponse(int status, String message, long timeStamp) {
         this.status = status;
         this.message = message;
         this.timeStamp = timeStamp;
