+++ src/main/java/com/teethcare/model/response/MessageResponse.java
@@ -1,0 +1,14 @@
+package com.teethcare.model.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class MessageResponse {
+    private String message;
+}
